Move as much as possible in lib and add a mechanism to retrive internals from a package require
