+++ lib/core/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+  Context: require('./Context'),
+  KuzzleProxy: require('./KuzzleProxy'),
+  pluginsInstall: require('./pluginsInstall')
+};
