+++ lib/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+  core: require('./core'),
+  service: require('./service'),
+  store: require('./store')
+};
