+++ lib/service/Broker.js
@@ -55,7 +55,7 @@ Broker.prototype.initiateServer = function (webSocketOptions) {
 /**
  * Triggered when a connection is made on the brokerSocket
  *
- * @param backendSocket
+ * @param {WebSocket} backendSocket
  */
 Broker.prototype.onConnection = function (backendSocket) {
   var
