+++ lib/service/index.js
@@ -1,0 +1,6 @@
+module.exports = {
+  Backend: require('./Backend'),
+  Broker: require('./Broker'),
+  HttpProxy: require('./HttpProxy'),
+  Router: require('./Router')
+};
