+++ lib/store/index.js
@@ -1,0 +1,6 @@
+module.exports = {
+  BackendConnection: require('./BackendConnection'),
+  ClientConnection: require('./ClientConnection'),
+  PendingRequest: require('./PendingRequest'),
+  Plugin: require('./Plugin')
+};
